fix: assume chainID is float64 (#211)
<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->

## Related Issue Or Context
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Otherwise, describe context and motivation for change herre -->

Closes: #<issue>

## How Has This Been Tested? Testing details.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have ensured that all acceptance criteria (or expected behavior)
from issue are met
- [ ] I have updated the documentation locally and in docs.
- [ ] I have added tests to cover my changes.
- [ ] I have ensured that all the checks are passing and green, I've
signed the CLA bot

+++ config/config.go
@@ -153,7 +153,7 @@ func MergeChainConfigs(chainConfigs []map[string]interface{}, solverConfig solve
 		}
 
 		cc := map[string]interface{}{
-			"id":       id,
+			"id":       float64(id),
 			"type":     "evm",
 			"name":     fmt.Sprintf("chain-%d", id),
 			"endpoint": c.Transport.HTTP[0],
@@ -166,8 +166,8 @@ func MergeChainConfigs(chainConfigs []map[string]interface{}, solverConfig solve
 
 func chainExists(chainConfigs []map[string]interface{}, chainID uint64) bool {
 	for _, cc := range chainConfigs {
-		ccID := cc["id"].(uint64)
-		if ccID == chainID {
+		ccID := cc["id"].(float64)
+		if uint64(ccID) == chainID {
 			return true
 		}
 	}
